requirements.txt
fastapi
uvicorn
pillow
pytesseract
requests
google-cloud-storage